+++ packages/uploadthing/src/fastify.ts
@@ -17,10 +17,9 @@ import {
 import { incompatibleNodeGuard } from "./internal/incompat-node-guard";
 import { initLogger } from "./internal/logger";
 import { toWebRequest } from "./internal/node-http/toWebRequest";
-import type { RouteHandlerOptions } from "./internal/types";
+import type { FileRouter, RouteHandlerOptions } from "./internal/types";
 import type { CreateBuilderOptions } from "./internal/upload-builder";
 import { createBuilder } from "./internal/upload-builder";
-import type { FileRouter } from "./types";
 
 export type { FileRouter };
 export { UTFiles } from "./internal/types";
