+++ packages/uploadthing/src/internal/types.ts
@@ -1,13 +1,22 @@
 import type {
   ContentDisposition,
+  ErrorMessage,
   FetchEsque,
+  FileRouterInputConfig,
   FileRouterInputKey,
   Json,
+  MaybePromise,
+  Simplify,
   UploadThingError,
 } from "@uploadthing/shared";
 
-import type { FileRouter, FileUploadData } from "../types";
+import type {
+  FileUploadData,
+  FileUploadDataWithCustomId,
+  UploadedFileData,
+} from "../types";
 import type { LogLevel } from "./logger";
+import type { JsonParser } from "./parser";
 
 interface UploadThingBaseResponse {
   key: string;
@@ -51,22 +60,31 @@ export type ActionType = (typeof VALID_ACTION_TYPES)[number];
  */
 export type UTEvents = {
   upload: {
-    files: FileUploadData[];
-    input: Json;
+    in: {
+      files: FileUploadData[];
+      input: Json;
+    };
+    out: PresignedURLs;
   };
   failure: {
-    fileKey: string;
-    uploadId: string | null;
-    s3Error?: string;
-    fileName: string;
+    in: {
+      fileKey: string;
+      uploadId: string | null;
+      s3Error?: string;
+      fileName: string;
+    };
+    out: null;
   };
   "multipart-complete": {
-    fileKey: string;
-    uploadId: string;
-    etags: {
-      tag: string;
-      partNumber: number;
-    }[];
+    in: {
+      fileKey: string;
+      uploadId: string;
+      etags: {
+        tag: string;
+        partNumber: number;
+      }[];
+    };
+    out: null;
   };
 };
 
@@ -135,5 +153,135 @@ export type RequestHandler<TArgs extends MiddlewareFnArgs<any, any, any>> = (
 ) => RequestHandlerOutput;
 
 /**
- * Inference helpers
+ * Builder types
  */
+
+const unsetMarker = "unsetMarker" as "unsetMarker" & {
+  __brand: "unsetMarker";
+};
+export type UnsetMarker = typeof unsetMarker;
+
+export type ValidMiddlewareObject = {
+  [UTFiles]?: FileUploadDataWithCustomId[];
+  [key: string]: unknown;
+};
+
+type ResolverOptions<TParams extends AnyParams> = {
+  metadata: Simplify<
+    TParams["_metadata"] extends UnsetMarker
+      ? undefined
+      : Omit<TParams["_metadata"], typeof UTFiles>
+  >;
+
+  file: UploadedFileData;
+};
+
+export interface AnyParams {
+  _input: any;
+  _metadata: any; // imaginary field used to bind metadata return type to an Upload resolver
+  _middlewareArgs: MiddlewareFnArgs<any, any, any>;
+  _errorShape: any;
+  _errorFn: any; // used for onUploadError
+  _output: any;
+}
+
+type MiddlewareFn<
+  TInput extends Json | UnsetMarker,
+  TOutput extends ValidMiddlewareObject,
+  TArgs extends MiddlewareFnArgs<any, any, any>,
+> = (
+  opts: TArgs & {
+    files: FileUploadData[];
+    input: TInput extends UnsetMarker ? undefined : TInput;
+  },
+) => MaybePromise<TOutput>;
+
+type ResolverFn<TOutput extends Json | void, TParams extends AnyParams> = (
+  opts: ResolverOptions<TParams>,
+) => MaybePromise<TOutput>;
+
+type UploadErrorFn = (input: {
+  error: UploadThingError;
+  fileKey: string;
+}) => void;
+
+export interface UploadBuilder<TParams extends AnyParams> {
+  input: <TParser extends JsonParser>(
+    parser: TParams["_input"] extends UnsetMarker
+      ? TParser
+      : ErrorMessage<"input is already set">,
+  ) => UploadBuilder<{
+    _input: TParser["_output"];
+    _metadata: TParams["_metadata"];
+    _middlewareArgs: TParams["_middlewareArgs"];
+    _errorShape: TParams["_errorShape"];
+    _errorFn: TParams["_errorFn"];
+    _output: UnsetMarker;
+  }>;
+  middleware: <TOutput extends ValidMiddlewareObject>(
+    fn: TParams["_metadata"] extends UnsetMarker
+      ? MiddlewareFn<TParams["_input"], TOutput, TParams["_middlewareArgs"]>
+      : ErrorMessage<"middleware is already set">,
+  ) => UploadBuilder<{
+    _input: TParams["_input"];
+    _metadata: TOutput;
+    _middlewareArgs: TParams["_middlewareArgs"];
+    _errorShape: TParams["_errorShape"];
+    _errorFn: TParams["_errorFn"];
+    _output: UnsetMarker;
+  }>;
+  onUploadComplete: <TOutput extends Json | void>(
+    fn: ResolverFn<TOutput, TParams>,
+  ) => Uploader<{
+    _input: TParams["_input"];
+    _metadata: TParams["_metadata"];
+    _middlewareArgs: TParams["_middlewareArgs"];
+    _errorShape: TParams["_errorShape"];
+    _errorFn: TParams["_errorFn"];
+    _output: TOutput;
+  }>;
+  onUploadError: (
+    fn: TParams["_errorFn"] extends UnsetMarker
+      ? UploadErrorFn
+      : ErrorMessage<"onUploadError is already set">,
+  ) => UploadBuilder<{
+    _input: TParams["_input"];
+    _metadata: TParams["_metadata"];
+    _middlewareArgs: TParams["_middlewareArgs"];
+    _errorShape: TParams["_errorShape"];
+    _errorFn: UploadErrorFn;
+    _output: UnsetMarker;
+  }>;
+}
+
+export type UploadBuilderDef<TParams extends AnyParams> = {
+  routerConfig: FileRouterInputConfig;
+  inputParser: JsonParser;
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  middleware: MiddlewareFn<TParams["_input"], {}, TParams["_middlewareArgs"]>;
+  errorFormatter: (err: UploadThingError) => TParams["_errorShape"];
+  onUploadError: UploadErrorFn;
+};
+
+export interface Uploader<TParams extends AnyParams> {
+  _def: TParams & UploadBuilderDef<TParams>;
+  resolver: ResolverFn<TParams["_output"], TParams>;
+}
+
+export type FileRouter<TParams extends AnyParams = AnyParams> = Record<
+  string,
+  Uploader<TParams>
+>;
+
+export type inferEndpointInput<TUploader extends Uploader<any>> =
+  TUploader["_def"]["_input"] extends UnsetMarker
+    ? undefined
+    : TUploader["_def"]["_input"];
+
+export type inferEndpointOutput<TUploader extends Uploader<any>> =
+  TUploader["_def"]["_output"] extends UnsetMarker | void | undefined
+    ? null
+    : TUploader["_def"]["_output"];
+
+export type inferErrorShape<TRouter extends FileRouter> =
+  TRouter[keyof TRouter]["_def"]["_errorShape"];
