+++ packages/uploadthing/test/gen-uploader.test.ts
@@ -2,9 +2,8 @@ import { describe, expectTypeOf, it } from "vitest";
 import { z } from "zod";
 
 import { genUploader } from "../src/client";
-import type { FileRouter } from "../src/internal/types";
 import { createBuilder } from "../src/internal/upload-builder";
-import type { UploadedFile } from "../src/types";
+import type { ClientUploadedFileData, FileRouter } from "../src/types";
 
 const doNotExecute = (_fn: (...args: any[]) => any) => {
   // noop
@@ -39,14 +38,14 @@ describe("genuploader", () => {
     doNotExecute(async () => {
       // @ts-expect-error - Input should be required here
       const res = await uploader("uploadable2", { files });
-      expectTypeOf<UploadedFile<{ baz: "qux" }>[]>(res);
+      expectTypeOf<ClientUploadedFileData<{ baz: "qux" }>[]>(res);
     });
   });
 
   it("types serverData as null if polling is skipped", () => {
     doNotExecute(async () => {
       const res = await uploader("uploadable1", { files, skipPolling: true });
-      expectTypeOf<UploadedFile<null>[]>(res);
+      expectTypeOf<ClientUploadedFileData<null>[]>(res);
     });
   });
 });
