+++ packages/uploadthing/test/request-handler.test.ts
@@ -67,7 +67,7 @@ describe("errors for invalid request input", () => {
         method: "POST",
         headers: baseHeaders,
         body: JSON.stringify({
-          files: [{ name: "foo.txt", size: 48 }],
+          files: [{ name: "foo.txt", size: 48, type: "text/plain" }],
         }),
       }),
     );
@@ -86,7 +86,7 @@ describe("errors for invalid request input", () => {
         method: "POST",
         headers: baseHeaders,
         body: JSON.stringify({
-          files: [{ name: "foo.txt", size: 48 }],
+          files: [{ name: "foo.txt", size: 48, type: "text/plain" }],
         }),
       }),
     );
@@ -108,7 +108,7 @@ describe("file route config", () => {
         method: "POST",
         headers: baseHeaders,
         body: JSON.stringify({
-          files: [{ name: "foo.txt", size: 48 }],
+          files: [{ name: "foo.txt", size: 48, type: "text/plain" }],
         }),
       }),
     );
@@ -127,7 +127,9 @@ describe("file route config", () => {
         method: "POST",
         headers: baseHeaders,
         body: JSON.stringify({
-          files: [{ name: "foo.png", size: 3 * 1024 * 1024 }],
+          files: [
+            { name: "foo.png", size: 3 * 1024 * 1024, type: "image/png" },
+          ],
         }),
       }),
     );
@@ -144,8 +146,8 @@ describe("file route config", () => {
         headers: baseHeaders,
         body: JSON.stringify({
           files: [
-            { name: "foo.png", size: 48 },
-            { name: "bar.png", size: 64 },
+            { name: "foo.png", size: 48, type: "image/png" },
+            { name: "bar.png", size: 64, type: "image/png" },
           ],
         }),
       }),
@@ -168,7 +170,7 @@ describe(".input()", () => {
         method: "POST",
         headers: baseHeaders,
         body: JSON.stringify({
-          files: [{ name: "foo.txt", size: 48 }],
+          files: [{ name: "foo.txt", size: 48, type: "text/plain" }],
         }),
       }),
     );
@@ -191,7 +193,7 @@ describe(".input()", () => {
         method: "POST",
         headers: baseHeaders,
         body: JSON.stringify({
-          files: [{ name: "foo.txt", size: 48 }],
+          files: [{ name: "foo.txt", size: 48, type: "text/plain" }],
           input: { foo: "QUX" },
         }),
       }),
@@ -217,7 +219,7 @@ describe(".input()", () => {
         method: "POST",
         headers: baseHeaders,
         body: JSON.stringify({
-          files: [{ name: "foo.txt", size: 48 }],
+          files: [{ name: "foo.txt", size: 48, type: "text/plain" }],
           input: { foo: "BAR" },
         }),
       }),
@@ -239,7 +241,7 @@ describe(".middleware()", () => {
         method: "POST",
         headers: baseHeaders,
         body: JSON.stringify({
-          files: [{ name: "foo.png", size: 48 }],
+          files: [{ name: "foo.png", size: 48, type: "image/png" }],
         }),
       }),
     );
@@ -247,7 +249,7 @@ describe(".middleware()", () => {
     expect(middlewareMock).toBeCalledTimes(1);
     expect(middlewareMock).toHaveBeenCalledWith(
       expect.objectContaining({
-        files: [{ name: "foo.png", size: 48 }],
+        files: [{ name: "foo.png", size: 48, type: "image/png" }],
       }),
     );
 
@@ -260,7 +262,7 @@ describe(".middleware()", () => {
         method: "POST",
         headers: baseHeaders,
         body: JSON.stringify({
-          files: [{ name: "foo.txt", size: 48 }],
+          files: [{ name: "foo.txt", size: 48, type: "text/plain" }],
         }),
       }),
     );
@@ -269,7 +271,7 @@ describe(".middleware()", () => {
     expect(middlewareMock).toHaveBeenCalledWith(
       expect.objectContaining({
         input: undefined,
-        files: [{ name: "foo.txt", size: 48 }],
+        files: [{ name: "foo.txt", size: 48, type: "text/plain" }],
       }),
     );
